market_intel_agent/requirements.txt
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
pyyaml>=6.0
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0
requests>=2.26.0
beautifulsoup4>=4.10.0
argparse>=1.4.0
python-dotenv>=0.19.0
tqdm>=4.62.0
yfinance>=0.2.12
newsapi-python>=0.2.6